// frontend/src/app/core/api/generated/model/updateProjectDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * TaskFlow API
 * Taskflow Backend com JWT + Prisma
 * OpenAPI spec version: 1.0
 */

export interface UpdateProjectDto {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** IDs dos usuários responsáveis associados ao projeto */
  userIds?: number[];
}